separated ORMs for messages and passages as well as MessageManager and PassageManager
