+++ letta/orm/file.py
@@ -27,4 +27,3 @@ class FileMetadata(SqlalchemyBase, OrganizationMixin, SourceMixin):
     # relationships
     organization: Mapped["Organization"] = relationship("Organization", back_populates="files", lazy="selectin")
     source: Mapped["Source"] = relationship("Source", back_populates="files", lazy="selectin")
-    passages: Mapped[List["Passage"]] = relationship("Passage", back_populates="file", lazy="selectin")
