+++ letta/orm/user.py
@@ -21,7 +21,6 @@ class User(SqlalchemyBase, OrganizationMixin):
     # relationships
     organization: Mapped["Organization"] = relationship("Organization", back_populates="users")
     messages: Mapped[List["Message"]] = relationship("Message", back_populates="user")
-    passages: Mapped[List["Passage"]] = relationship("Passage", back_populates="user")
 
     # TODO: Add this back later potentially
     # agents: Mapped[List["Agent"]] = relationship(
